requirements.txt
streamlit==1.40.1
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
